Move IpAddr and SocketAddr to core
